+++ lib/database_consistency/writers/base_writer.rb
@@ -4,15 +4,16 @@ module DatabaseConsistency
   module Writers
     # The base class for writers
     class BaseWriter
-      attr_reader :results, :config
+      attr_reader :results, :config, :opts
 
-      def initialize(results, config: Configuration.new)
+      def initialize(results, config: Configuration.new, opts: nil)
         @results = results
         @config = config
+        @opts = opts
       end
 
-      def self.write(results, config: Configuration.new)
-        new(results, config: config).write
+      def self.write(results, config: Configuration.new, opts: nil)
+        new(results, config: config, opts: opts).write
       end
     end
   end